edit puthex unsigned int

+++ ft_putupperhex_fd.c
@@ -6,13 +6,13 @@
 /*   By: skawanis <skawanis@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/27 19:20:42 by skawanis          #+#    #+#             */
-/*   Updated: 2023/03/27 20:35:59 by skawanis         ###   ########.fr       */
+/*   Updated: 2023/03/27 22:23:20 by skawanis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static void	rec(size_t n, int fd, size_t *count)
+static void	rec(unsigned int n, int fd, size_t *count)
 {
 	if (n > 15)
 	{
@@ -21,7 +21,7 @@ static void	rec(size_t n, int fd, size_t *count)
 	*count += write(fd, &"0123456789ABCDEF"[n % 16], 1);
 }
 
-ssize_t	ft_putupperhex_fd(size_t n, int fd)
+ssize_t	ft_putupperhex_fd(unsigned int n, int fd)
 {
 	size_t	count;
